fixed delete methods

+++ src/modules/projects/projects.service.ts
@@ -173,9 +173,13 @@ export class ProjectsService {
     }
     await this.voteModel.deleteMany({ project: deletedProject._id }).exec();
     await this.reviewModel.deleteMany({ project: deletedProject._id }).exec();
-    await this.uploadService.remove(deletedProject.logo.split('/').slice(-1)[0]);
-    for (const screenshot of deletedProject.screenshots) {
-      await this.uploadService.remove(screenshot.split('/').slice(-1)[0]);
+    if (deletedProject.logo) {
+      await this.uploadService.remove(deletedProject.logo.split('/').slice(-1)[0]);
+    }
+    if (deletedProject.screenshots.length > 0) {
+      for (const screenshot of deletedProject.screenshots) {
+        await this.uploadService.remove(screenshot.split('/').slice(-1)[0]);
+      }
     }
     return { ...deletedProject.toObject(), voted: false };
   }
